Factor out common logic
